+++ antenna/miniprogram/app.js
@@ -1,5 +1,5 @@
 "use strict";
-const APP_VERSION = "0.10.0";
+const APP_VERSION = "0.10.1";
 App({
     globalData: {
         appVersion: APP_VERSION,